Moved table-logic out of hook
