+++ mottak-arkiv-web/src/components/AllArkivuttrekk.tsx
@@ -1,17 +1,14 @@
 import React from 'react'
-import ArkivuttrekkTable from './ArkivuttrekkTable'
+import ArkivuttrekkTable from './arkivuttrekk-table/ArkivuttrekkTable'
 
 /**
  * Display all arkivuttrekk as table data.
  */
-const AllArkivuttrekk: React.FC = ():JSX.Element => {
-
-	return (
-		<>
-			<h2>Alle arkivuttrekk</h2>
-			<ArkivuttrekkTable pagination={false}/>
-		</>
-	)
-}
+const AllArkivuttrekk: React.FC = ():JSX.Element => (
+	<>
+		<h2>Alle arkivuttrekk</h2>
+		<ArkivuttrekkTable pagination={false}/>
+	</>
+)
 
 export default AllArkivuttrekk
