+++ mottak-arkiv-web/src/components/arkivuttrekk-table/ArkivuttrekkRow.tsx
@@ -11,7 +11,8 @@ import {
 } from '@material-ui/core'
 import { Link } from 'react-router-dom'
 import { makeStyles } from '@material-ui/core/styles'
-import { ArkivUttrekk } from '../types/sharedTypes'
+
+import { ArkivUttrekk } from '../../types/sharedTypes'
 
 const useStyles = makeStyles(theme => ({
 	root: {
